+++ CMakeLists.txt
@@ -8,10 +8,13 @@ find_package(CURL REQUIRED)
 find_package(PkgConfig REQUIRED)
 pkg_check_modules(Xlsxwriter REQUIRED IMPORTED_TARGET xlsxwriter)
 
+include_directories(src/api)
+
 # Add the executable
-add_executable(trading_bot test/main.cpp binance_api.cpp
-        binance_api.cpp
-        test/trade_endpoint.cpp)
+add_executable(trading_bot test/main.cpp src/api/binance/binance_api.cpp
+        src/api/binance/binance_api.cpp
+        test/trade_endpoint.cpp
+        src/api/curl_main.cpp)
 
 target_link_libraries(${PROJECT_NAME} PkgConfig::Xlsxwriter)
 target_link_libraries(trading_bot CURL::libcurl)
