+++ src/api/binance/binance_api.h
@@ -6,11 +6,12 @@
 #define BINANCE_API_H
 
 #include <string>
+#include "curl_main.h"
 
 class BinanceAPI {
 public:
-    BinanceAPI();
-    ~BinanceAPI();
+    BinanceAPI() = default;
+
     static std::string ping();
     static std::string exchangeInfo();
     static std::string depth(const std::string &symbol);
@@ -22,10 +23,8 @@ public:
 //    static std::string fetchVolume(const std::string& url);
 //    static std:: void trade (const std::string& url);
 
-
 private:
-    static std::string callAPI(const std::string &endpoint);
-    static size_t WriteCallback(void *contents, size_t size, size_t nmemb, std::string *data);
+    static CurlMain curlMain;
 
 protected:
