+++ test/main.cpp
@@ -7,7 +7,7 @@
 #include <filesystem> // For filesystem operations (C++17)
 #include <xlsxwriter.h>
 #include <sstream>
-#include "../binance_api.h"
+#include "../src/api/binance/binance_api.h"
 
 namespace fs = std::filesystem;
